regrouped functions for geom
