+++ scripts/conformer/geom_stats.py
@@ -9,70 +9,17 @@ from rdkit import Chem
 from pathlib import Path
 from tqdm import tqdm
 
-from gflownet.utils.molecule.rotatable_bonds import get_rotatable_ta_list, is_hydrogen_ta
+from gflownet.utils.molecule.rotatable_bonds import get_rotatable_ta_list, has_hydrogen_tas
+from gflownet.utils.molecule.geom import get_conf_geom,  get_all_confs_geom, get_rd_mol, all_same_graphs
 
 """
 Here we use rdkit_folder format of the GEOM dataset 
 Tutorial and downloading links are here: https://github.com/learningmatter-mit/geom/tree/master
 """
-
-def get_conf_geom(base_path, smiles, conf_idx=0, summary_file=None):
-    if summary_file is None:
-        drugs_file = base_path / 'rdkit_folder/summary_drugs.json'
-        with open(drugs_file, "r") as f:
-            summary_file = json.load(f)
-
-    pickle_path = base_path / "rdkit_folder" / summary_file[smiles]['pickle_path']
-    if os.path.isfile(pickle_path):
-        with open(pickle_path, "rb") as f:
-            dic = pickle.load(f)
-        mol = dic['conformers'][conf_idx]['rd_mol']
-        return mol
-
-def get_all_confs_geom(base_path, smiles, summary_file=None):
-    if summary_file is None:
-        drugs_file = base_path / 'rdkit_folder/summary_drugs.json'
-        with open(drugs_file, "r") as f:
-            summary_file = json.load(f)
-    try:
-        pickle_path = base_path / "rdkit_folder" / summary_file[smiles]['pickle_path']
-        if os.path.isfile(pickle_path):
-            with open(pickle_path, "rb") as f:
-                dic = pickle.load(f)
-            conformers = [x['rd_mol'] for x in dic['conformers']]
-            return conformers
-    except KeyError:
-        print('No pickle_path file for {}'.format(smiles))
-
-def get_rd_mol(smiles):
-    mol = Chem.MolFromSmiles(smiles)
-    mol = Chem.AddHs(mol)
-    return mol
-
-def has_same_can_smiles(mol1, mol2):
-    sm1 = Chem.CanonSmiles(Chem.MolToSmiles(mol1))
-    sm2 = Chem.CanonSmiles(Chem.MolToSmiles(mol2))
-    return sm1 == sm2
-
-def all_same_graphs(mols):
-    ref = mols[0]
-    same = []
-    for mol in mols:
-        same.append(has_same_can_smiles(ref, mol))
-    return np.all(same)
-
-def has_hydrogen_tas(mol):
-    tas = get_rotatable_ta_list(mol)
-    hydrogen_flags = []
-    for t in tas:
-        hydrogen_flags.append(is_hydrogen_ta(mol, t))
-    return np.any(hydrogen_flags)
-
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--geom_dir', type=str, default='/home/mila/a/alexandra.volokhova/scratch/datasets/geom')
-    parser.add_argument('--output_dir', type=str, default='./')
+    parser.add_argument('--output_file', type=str, default='./geom_stats.csv')
     args = parser.parse_args()
 
     base_path = Path(args.geom_dir)
@@ -116,6 +63,6 @@ if __name__ == '__main__':
     'n_atoms': n_atoms,
     }
     df = pd.DataFrame(data)
-    df.to_csv(os.path.join(args.output_dir, 'geom_stats.csv'))
+    df.to_csv(args.output_file)
